refactor(workers): improve Sentry sync extensibility
